Clean up `fractal_tasks_non_canonical` fake package
