--- tests/v2/fractal_tasks_mock_capital/README.md
@@ -1,9 +1,0 @@
-Prepare manifest
-```bash
-poetry run -C ../../../ python init_manifest.py
-```
-
-Basic build
-```bash
-python -m build
-```
