--- tests/v2/fractal_tasks_mock_capital/src/FrAcTaL_TaSkS_MoCk_NoN_CaNoNiCaL.egg-info/SOURCES.txt
@@ -1,29 +1,0 @@
-README.md
-pyproject.toml
-src/FrAcTaL_TaSkS_MoCk_NoN_CaNoNiCaL.egg-info/PKG-INFO
-src/FrAcTaL_TaSkS_MoCk_NoN_CaNoNiCaL.egg-info/SOURCES.txt
-src/FrAcTaL_TaSkS_MoCk_NoN_CaNoNiCaL.egg-info/dependency_links.txt
-src/FrAcTaL_TaSkS_MoCk_NoN_CaNoNiCaL.egg-info/requires.txt
-src/FrAcTaL_TaSkS_MoCk_NoN_CaNoNiCaL.egg-info/top_level.txt
-src/fractal_tasks_mock/__FRACTAL_MANIFEST__.json
-src/fractal_tasks_mock/__init__.py
-src/fractal_tasks_mock/apply_registration_to_image.py
-src/fractal_tasks_mock/calculate_registration_compute.py
-src/fractal_tasks_mock/calculate_registration_init.py
-src/fractal_tasks_mock/cellpose_segmentation.py
-src/fractal_tasks_mock/create_cellvoyager_ome_zarr.py
-src/fractal_tasks_mock/create_cellvoyager_ome_zarr_multiplex.py
-src/fractal_tasks_mock/dummy_insert_single_image.py
-src/fractal_tasks_mock/dummy_remove_images.py
-src/fractal_tasks_mock/dummy_unset_attribute.py
-src/fractal_tasks_mock/fill_cellvoyager_ome_zarr.py
-src/fractal_tasks_mock/find_registration_consensus.py
-src/fractal_tasks_mock/generic_task.py
-src/fractal_tasks_mock/generic_task_parallel.py
-src/fractal_tasks_mock/illumination_correction.py
-src/fractal_tasks_mock/illumination_correction_compute.py
-src/fractal_tasks_mock/illumination_correction_init.py
-src/fractal_tasks_mock/input_models.py
-src/fractal_tasks_mock/maximum_intensity_projection.py
-src/fractal_tasks_mock/new_ome_zarr.py
-src/fractal_tasks_mock/utils.py
