--- tests/v2/fractal_tasks_mock_capital/src/FrAcTaL_TaSkS_MoCk_NoN_CaNoNiCaL.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-fractal_tasks_mock
