--- tests/v2/fractal_tasks_mock_capital/src/fractal_tasks_mock/calculate_registration_compute.py
@@ -1,43 +1,0 @@
-from pathlib import Path
-
-from fractal_tasks_mock.input_models import InitArgsRegistration
-from pydantic.decorator import validate_arguments
-
-
-@validate_arguments
-def calculate_registration_compute(
-    *,
-    zarr_url: str,
-    init_args: InitArgsRegistration,
-) -> None:
-    """
-    Dummy task description.
-
-    Arguments:
-        path: description
-        init_args: description
-    """
-
-    ref_zarr_url = init_args.ref_zarr_url
-    print("[calculate_registration_compute] START")
-    print(f"[calculate_registration_compute] {zarr_url=}")
-    print(f"[calculate_registration_compute] {ref_zarr_url=}")
-
-    table_path = Path(zarr_url) / "registration_table"
-    print(
-        f"[calculate_registration_compute] Writing to {table_path.as_posix()}"
-    )
-
-    with table_path.open("w") as f:
-        f.write(
-            "Calculate registration for\n"
-            f"{zarr_url=}\n"
-            f"{ref_zarr_url=}\n"
-        )
-    print("[calculate_registration_compute] END")
-
-
-if __name__ == "__main__":
-    from utils import run_fractal_task
-
-    run_fractal_task(task_function=calculate_registration_compute)
