--- tests/v2/fractal_tasks_mock_capital/src/fractal_tasks_mock/create_cellvoyager_ome_zarr.py
@@ -1,74 +1,0 @@
-import logging
-from pathlib import Path
-
-from pydantic.decorator import validate_arguments
-
-
-@validate_arguments
-def create_cellvoyager_ome_zarr(
-    *,
-    zarr_urls: list[str],
-    zarr_dir: str,
-    image_dir: str,
-    num_images: int = 2,
-) -> dict:
-    """
-    Dummy task description.
-
-    Arguments:
-        zarr_urls: description
-        zarr_dir: description
-        image_dir: Image where the raw images are
-        num_images: Number of images that this dummy task will produce.
-    """
-
-    zarr_dir = zarr_dir.rstrip("/")
-
-    if len(zarr_urls) > 0:
-        raise ValueError(
-            "Error in create_cellvoyager_ome_zarr: "
-            "`zarr_urls` argument must be empty, but "
-            f"{zarr_urls=}."
-        )
-
-    # Based on images in image_folder, create plate OME-Zarr
-    Path(zarr_dir).mkdir(parents=True)
-    plate_zarr_name = "my_plate.zarr"
-    zarr_path = (Path(zarr_dir) / plate_zarr_name).as_posix()
-
-    logging.info("[create_cellvoyager_ome_zarr] START")
-    logging.info(f"[create_cellvoyager_ome_zarr] {image_dir=}")
-    logging.info(f"[create_cellvoyager_ome_zarr] {zarr_dir=}")
-    logging.info(f"[create_cellvoyager_ome_zarr] {zarr_path=}")
-
-    # Create (fake) OME-Zarr folder on disk
-    Path(zarr_path).mkdir()
-
-    # Prepare fake list of OME-Zarr images
-    image_relative_paths = [
-        f"A/{ind_image:02d}/0" for ind_image in range(1, num_images + 1)
-    ]
-
-    # Create well/image OME-Zarr folders on disk, and prepare output
-    # metadata
-    parallelization_list = []
-    for image_relative_path in image_relative_paths:
-        (Path(zarr_path) / image_relative_path).mkdir(parents=True)
-        zarr_url = f"{zarr_dir}/{plate_zarr_name}/{image_relative_path}"
-        raw_zarr_url = (
-            Path(image_dir) / image_relative_path.replace("/", "_")
-        ).as_posix() + ".tif"
-        parallelization_list.append(
-            dict(
-                zarr_url=zarr_url,
-                init_args=dict(raw_zarr_url=raw_zarr_url),
-            )
-        )
-    logging.info("[create_cellvoyager_ome_zarr] END")
-    return dict(parallelization_list=parallelization_list)
-
-
-if __name__ == "__main__":
-    from utils import run_fractal_task
-
-    run_fractal_task(task_function=create_cellvoyager_ome_zarr)
