--- tests/v2/fractal_tasks_mock_capital/src/fractal_tasks_mock/new_ome_zarr.py
@@ -1,59 +1,0 @@
-import logging
-from pathlib import Path
-
-from fractal_tasks_mock.utils import _extract_common_root
-from pydantic.decorator import validate_arguments
-
-
-@validate_arguments
-def new_ome_zarr(
-    *,
-    zarr_urls: list[str],
-    zarr_dir: str,
-    suffix: str = "new",
-) -> dict:
-    """
-    Dummy task description.
-    """
-
-    dict_shared = _extract_common_root(zarr_urls)
-    old_plate = dict_shared.get("shared_plate")
-
-    logging.info("[new_ome_zarr] START")
-    logging.info(f"[new_ome_zarr] {zarr_urls=}")
-    logging.info(f"[new_ome_zarr] Identified {old_plate=}")
-
-    assert old_plate.endswith(".zarr")  # nosec
-    new_plate = old_plate.strip(".zarr") + f"_{suffix}.zarr"
-    logging.info(f"[new_ome_zarr] {new_plate=}")
-
-    # Based on images in image_folder, create plate OME-Zarr
-    new_zarr_path = (Path(zarr_dir) / new_plate).as_posix()
-    logging.info(f"[new_ome_zarr] {new_zarr_path=}")
-
-    # Create (fake) OME-Zarr folder on disk
-    Path(new_zarr_path).mkdir()
-
-    parallelization_list = []
-    for old_zarr_url in zarr_urls:
-        new_zarr_url = old_zarr_url.replace(old_plate, new_plate)
-        parallelization_list.append(
-            dict(
-                zarr_url=old_zarr_url,
-                init_args=dict(
-                    new_zarr_url=new_zarr_url,
-                    new_plate=new_plate,
-                ),
-            )
-        )
-
-    # Prepare output metadata
-    out = dict(parallelization_list=parallelization_list)
-    logging.info("[new_ome_zarr] END")
-    return out
-
-
-if __name__ == "__main__":
-    from utils import run_fractal_task
-
-    run_fractal_task(task_function=new_ome_zarr)
