+++ tests/v2/fractal_tasks_non_canonical/src/FrAcTaL_TaSkS_MoCk_NoN_CaNoNiCaL.egg-info/PKG-INFO
@@ -1,7 +1,7 @@
 Metadata-Version: 2.1
 Name: FrAcTaL-TaSkS-MoCk-NoN-CaNoNiCaL
 Version: 0.0.1
-Summary: Fake tasks package
+Summary: Fake tasks package with non-canonical project name
 Author-email: Your Name <first.last@example.com>
 License: BSD-3-Clause
 Requires-Python: >=3.8
@@ -12,13 +12,3 @@ Requires-Dist: fractal-tasks-core; extra == "dev"
 Requires-Dist: build; extra == "dev"
 Provides-Extra: my-extra
 Requires-Dist: devtools; extra == "my-extra"
-
-Prepare manifest
-```bash
-poetry run -C ../../../ python init_manifest.py
-```
-
-Basic build
-```bash
-python -m build
-```
