+++ tests/v2/fractal_tasks_non_canonical/src/FrAcTaL_TaSkS_MoCk_NoN_CaNoNiCaL.egg-info/requires.txt
