+++ tests/v2/fractal_tasks_non_canonical/src/FrAcTaL_TaSkS_NoN_CaNoNiCaL.egg-info/requires.txt
@@ -1,0 +1,8 @@
+pydantic<2
+
+[dev]
+fractal-tasks-core
+build
+
+[my_extra]
+devtools
