+++ tests/v2/fractal_tasks_non_canonical/src/fractal_tasks_non_canonical/__FRACTAL_MANIFEST__.json
@@ -1,0 +1,76 @@
+{
+  "args_schema_version": "pydantic_v1",
+  "has_args_schemas": true,
+  "manifest_version": "2",
+  "task_list": [
+    {
+      "args_schema_non_parallel": {
+        "additionalProperties": false,
+        "properties": {
+          "raise_error": {
+            "default": false,
+            "description": "If true, the task raises a ValueError",
+            "title": "Raise Error",
+            "type": "boolean"
+          },
+          "sleep_time": {
+            "default": 1,
+            "description": "Time to sleep, in seconds",
+            "title": "Sleep Time",
+            "type": "integer"
+          },
+          "zarr_dir": {
+            "description": "description",
+            "title": "Zarr Dir",
+            "type": "string"
+          },
+          "zarr_urls": {
+            "description": "description",
+            "items": {
+              "type": "string"
+            },
+            "title": "Zarr Urls",
+            "type": "array"
+          }
+        },
+        "required": [
+          "zarr_urls",
+          "zarr_dir"
+        ],
+        "title": "GenericTask",
+        "type": "object"
+      },
+      "docs_info": "This is task generic_task.",
+      "docs_link": "https://example.org",
+      "executable_non_parallel": "generic_task.py",
+      "name": "generic_task"
+    },
+    {
+      "args_schema_parallel": {
+        "additionalProperties": false,
+        "properties": {
+          "zarr_url": {
+            "description": "Missing description",
+            "title": "Zarr Url",
+            "type": "string"
+          }
+        },
+        "required": [
+          "zarr_url"
+        ],
+        "title": "GenericTaskParallel",
+        "type": "object"
+      },
+      "docs_info": "This is task generic_task_parallel.",
+      "docs_link": "https://example.org",
+      "executable_parallel": "generic_task_parallel.py",
+      "input_types": {
+        "my_type": false
+      },
+      "name": "generic_task_parallel",
+      "output_types": {
+        "my_type": true
+      }
+    }
+  ]
+}
