+++ tests/v2/fractal_tasks_non_canonical/src/fractal_tasks_non_canonical/generic_task_parallel.py
