+++ tests/v2/fractal_tasks_non_canonical/update_package_manifest_and_wheel.sh
@@ -6,7 +6,7 @@ if [ ! -d "venv" ]; then
     python3 -m venv venv
 fi
 source venv/bin/activate
-python -m pip install -e ".[dev,my_extra]"
-python create_manifest.py
-python -m build
+python3 -m pip install -e ".[dev]"
+python3 create_manifest.py
+python3 -m build
 deactivate
